requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions

pip~=22.0.3
wheel~=0.36.2
setuptools~=57.0.0
cryptography~=36.0.1
msal~=1.17.0
six~=1.16.0
requests~=2.27.1
urllib3~=1.26.8
msrest~=0.6.21
isodate~=0.6.1
pywin32~=303
oauthlib~=3.2.0
certifi~=2021.10.8
idna~=3.3
pycparser~=2.21
portalocker~=2.4.0